# apps/api/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: LEOPARD API
  description: |
    LEOPARD delivery platform — REST API for customers, drivers, fleet owners, and administrators.
  version: "0.0.0"
  contact:
    name: LEOPARD Team
servers:
  - url: /api/v1
    description: Base path (routed through API gateway / reverse proxy)
security:
  - bearerAuth: []
paths:
...
  /orders:
    post:
      operationId: orders.create
      summary: Create a new order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
    get:
      operationId: orders.list
      summary: List own orders
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/OrderStatusFilter"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Paginated orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrderSchema"
        "401":
          $ref: "#/components/responses/ApiError"
  /orders/{id}:
    get:
      operationId: orders.getById
      summary: Order detail
      description: Accessible by owner, assigned driver, fleet owner, or admin.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
  /orders/{id}/cancel:
    post:
      operationId: orders.cancel
      summary: Cancel an order
      description: Only owner or admin may cancel, subject to cancellation rules.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderRequest"
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
  /orders/{id}/tracking:
    get:
      operationId: orders.getTracking
      summary: Paginated order tracking points
      description: Accessible by owner, assigned driver, fleet owner, or admin.
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Tracking points ordered by capturedAt DESC, id DESC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingPointPage"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
  # ── Maps ──────────────────────────────────────────────────────────────
...
  /fleet/profile:
    get:
      operationId: fleet.getProfile
      summary: Fleet owner profile
      tags: [Fleet]
      responses:
        "200":
          description: Fleet profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FleetProfile"
        "401":
          $ref: "#/components/responses/ApiError"
  /fleet/drivers:
    get:
      operationId: fleet.listDrivers
      summary: List drivers in the fleet
      tags: [Fleet]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FleetMemberStatus"
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Fleet drivers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/FleetDriver"
        "401":
          $ref: "#/components/responses/ApiError"
  /fleet/orders:
    get:
      operationId: fleet.listOrders
      summary: List orders assigned to fleet drivers
      tags: [Fleet]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: driverId
          in: query
          schema:
            type: string
          description: Filter by driver
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start of date range
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End of date range
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Fleet orders
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationEnvelope"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrderSchema"
        "401":
          $ref: "#/components/responses/ApiError"
  /fleet/orders/{id}:
    get:
      operationId: fleet.getOrderById
      summary: Fleet order detail
      tags: [Fleet]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSchema"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
  /fleet/orders/{id}/tracking:
    get:
      operationId: fleet.getOrderTracking
      summary: Paginated fleet order tracking points
      tags: [Fleet]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
        - $ref: "#/components/parameters/FromParam"
        - $ref: "#/components/parameters/ToParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Tracking points ordered by capturedAt DESC, id DESC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingPointPage"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
  # ── Media & Payment ───────────────────────────────────────────────────
  /orders/{id}/media/cargo:
    post:
      operationId: media.uploadCargo
      summary: Upload cargo image
      tags: [Media]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaUploadRequest"
      responses:
        "201":
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaRecord"
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
...
  /media/{id}/url:
    get:
      operationId: media.getSignedUrl
      summary: Signed read URL for media
      tags: [Media]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Media record ID
      responses:
        "200":
          description: Signed URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedUrlResponse"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
  /orders/{id}/payments:
    post:
      operationId: payments.createIntent
      summary: Create QR payment intent
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentRequest"
      responses:
        "201":
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentQrDto"
        "400":
          $ref: "#/components/responses/ApiError"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
    get:
      operationId: payments.getHistory
      summary: Payment history for an order
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/OrderIdParam"
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRecord"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
...
  /invoices/order/{orderId}:
    get:
      operationId: invoices.getForOrder
      summary: Invoice metadata for an order
      tags: [Invoices]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
          description: Order ID
      responses:
        "200":
          description: Invoice metadata and a short-lived view URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceView"
        "401":
          $ref: "#/components/responses/ApiError"
        "404":
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase-issued or session JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    OrderStatusFilter:
      name: status
      in: query
      schema:
        $ref: "#/components/schemas/OrderStatus"
      description: Filter by order status
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Sort expression (field allow-list enforced server-side)
    OrderIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Order ID
    FromParam:
      name: from
      in: query
      schema:
        type: string
        format: date-time
        pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d{1,3})?Z$'
      description: Inclusive lower capturedAt bound as an ISO 8601 UTC timestamp
    ToParam:
      name: to
      in: query
      schema:
        type: string
        format: date-time
        pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d{1,3})?Z$'
      description: Inclusive upper capturedAt bound as an ISO 8601 UTC timestamp
  responses:
    ApiError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorEnvelope"
  schemas:
...
    FleetMemberRole:
      type: string
      enum: [OWNER, DRIVER]
    FleetMemberStatus:
      type: string
      enum: [INVITED, ACTIVE, REMOVED]
    DriverAvailability:
      type: string
      enum: [OFFLINE, AVAILABLE, BUSY]
    VehicleType:
      type: string
      enum: [MOTORBIKE, VAN, TRUCK]
    OrderStatus:
      type: string
      enum: [REQUESTED, ACCEPTED, PICKING_UP, IN_TRANSIT, DELIVERED, CANCELLED]
    StopType:
      type: string
      enum: [PICKUP, STOP, DROPOFF]
    MediaType:
      type: string
      enum: [CARGO, DELIVERY_PROOF]
    PaymentStatus:
      type: string
      enum: [UNPAID, QR_CREATED, PAID_MANUAL, FAILED]
...
    PaymentProvider:
      type: string
      enum: [DEMO, PAYOS, VIETQR]
...
    PaginationEnvelope:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items: {}
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ApiErrorEnvelope:
      type: object
      required: [statusCode, code, message, requestId, timestamp]
      properties:
        statusCode:
          type: integer
        code:
          type: string
          description: Machine-readable error code (e.g. ORDER_NOT_FOUND)
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Correlation ID for tracing
        timestamp:
          type: string
          format: date-time
        details:
          description: Optional structured error details (validation errors, etc.)

    # ══ Health ══
...
    EstopSchema:
      type: object
      required: [type, address, lat, lng]
      properties:
        type:
          $ref: "#/components/schemas/StopType"
        address:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        contactName:
          type: string
        contactPhone:
          type: string
        note:
          type: string
...
    CreateOrderRequest:
      type: object
      required: [pickup, dropoff, vehicleType, estimateToken]
      properties:
        pickup:
          $ref: "#/components/schemas/EstopSchema"
        stops:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/EstopSchema"
        dropoff:
          $ref: "#/components/schemas/EstopSchema"
        vehicleType:
          $ref: "#/components/schemas/VehicleType"
        estimateToken:
          type: string
        cargoNote:
          type: string
        cargoWeightKg:
          type: number
    CancelOrderRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          description: Cancellation reason
    OrderSchema:
      type: object
      required: [id, status, vehicleType, pickup, dropoff, totalPriceVnd, createdAt]
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        vehicleType:
          $ref: "#/components/schemas/VehicleType"
        pickup:
          $ref: "#/components/schemas/EstopSchema"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/EstopSchema"
        dropoff:
          $ref: "#/components/schemas/EstopSchema"
        totalPriceVnd:
          type: integer
        distanceM:
          type: integer
        durationS:
          type: integer
        polyline:
          type: string
        cargoNote:
          type: string
        cargoWeightKg:
          type: number
        customerId:
          type: string
        driverId:
          type: string
        fleetId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        cancelReason:
          type: string

    # ══ Tracking ══
    TrackingPoint:
      type: object
      required:
        [id, orderId, driverId, clientPointId, latitude, longitude, capturedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        driverId:
          type: string
          format: uuid
        clientPointId:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        accuracyM:
          type: number
          format: double
          minimum: 0
        capturedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TrackingPointPage:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrackingPoint"
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    FleetProfile:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        logoUrl:
          type: string
          format: uri
        driverCount:
          type: integer
        activeOrderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FleetDriver:
      type: object
      required: [id, name, status, availability, joinedAt]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        status:
          $ref: "#/components/schemas/FleetMemberStatus"
        role:
          $ref: "#/components/schemas/FleetMemberRole"
        availability:
          $ref: "#/components/schemas/DriverAvailability"
        activeOrderCount:
          type: integer
        joinedAt:
          type: string
          format: date-time

    # ══ Media ══
    MediaUploadRequest:
      type: object
      required: [file, clientRequestId]
      properties:
        file:
          type: string
          format: binary
          description: Image file (JPEG, PNG, WEBP)
        clientRequestId:
          type: string
          format: uuid
          description: Idempotency key for this upload
    MediaRecord:
      type: object
      additionalProperties: false
      required: [id, type, orderId, uploadedBy, createdAt]
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/MediaType"
        orderId:
          type: string
        uploadedBy:
          type: string
        url:
          type: string
          format: uri
          description: Pre-signed read URL (may expire)
        createdAt:
          type: string
          format: date-time
    SignedUrlResponse:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time

    # ══ Payments ══
    CreatePaymentIntentRequest:
      type: object
      additionalProperties: false
      required: [clientRequestId]
      properties:
        clientRequestId:
          type: string
          format: uuid
    PaymentQrDto:
      type: object
      additionalProperties: false
      required: [amountVnd, provider, providerReference, expiresAt, qrPayload]
      properties:
        amountVnd:
          type: integer
          minimum: 0
        provider:
          $ref: "#/components/schemas/PaymentProvider"
        providerReference:
          type: string
        expiresAt:
          type: string
          format: date-time
        qrPayload:
          type: string
          description: Provider-neutral QR representation; never a raw provider response
    PaymentRecord:
      type: object
      additionalProperties: false
      required: [id, orderId, amountVnd, status, createdAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        amountVnd:
          type: integer
        status:
          $ref: "#/components/schemas/PaymentStatus"
        provider:
          $ref: "#/components/schemas/PaymentProvider"
        providerReference:
          type: string
        expiresAt:
          type: string
          format: date-time
        qrPayload:
          type: string
          description: Provider-neutral QR representation; never a raw provider response
        confirmedBy:
          type: string
        confirmedAt:
          type: string
          format: date-time
        note:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InvoiceView:
      type: object
      additionalProperties: false
      required: [id, invoiceNumber, amountVnd, vatRateVnd, totalVnd, issuedAt, emailSentAt, viewUrl]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
          description: 'Internal sequential number, format LP/YYYY/NNNNNN'
        amountVnd:
          type: integer
          minimum: 0
        vatRateVnd:
          type: integer
          minimum: 0
        totalVnd:
          type: integer
          minimum: 0
        issuedAt:
          type: string
          format: date-time
        emailSentAt:
          type: string
          format: date-time
          nullable: true
        viewUrl:
          type: string
          description: Short-lived, signed URL to view/download the PDF — never a raw storage key
...